compress.c: fix skips/distortion introduced in r7146
We need to ensure we're working with signed types when assigning
them.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7181 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/compress.c
@@ -354,7 +354,7 @@ void CompressDo(void *data, unsigned int length)
 	if (!pos)
 		pos = 1;
 
-	gr = ((gainTarget - gainCurrent) << 16)/pos;
+	gr = ((gainTarget - gainCurrent) << 16)/(int)pos;
 
 	/* Do the shiznit */
 	gf = gainCurrent << 16;
